Prepare for pub publishing.
So we can use it and continue the migration from the Cocoon side.
