+++ dev/benchmarks/metrics_center/lib/metrics_center.dart
@@ -1,0 +1,4 @@
+export 'src/common.dart';
+export 'src/flutter.dart';
+export 'src/google_benchmark.dart';
+export 'src/skiaperf.dart';
