+++ dev/benchmarks/metrics_center/lib/src/gcs_lock.dart
@@ -1,3 +1,7 @@
+// Copyright 2014 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
 import 'package:googleapis/storage/v1.dart';
 import 'package:googleapis_auth/auth.dart';
