+++ dev/benchmarks/metrics_center/pubspec.yaml
@@ -1,4 +1,9 @@
 name: metrics_center
+version: 0.0.2
+description:
+  Support multiple performance metrics sources/formats and destinations.
+homepage:
+  https://github.com/flutter/flutter/tree/master/dev/benchmarks/metrics_center
 
 environment:
   sdk: '>=2.10.0 <3.0.0'
@@ -11,8 +16,6 @@ dependencies:
   googleapis_auth: 0.2.12
   github: 7.0.4
   equatable: 1.2.5
-  mockito: 4.1.1
-  fake_async: 1.1.0
 
   _discoveryapis_commons: 0.2.0 # THIS LINE IS AUTOGENERATED - TO UPDATE USE "flutter update-packages --force-upgrade"
   async: 2.5.0-nullsafety.3 # THIS LINE IS AUTOGENERATED - TO UPDATE USE "flutter update-packages --force-upgrade"
@@ -38,6 +41,8 @@ dependencies:
 dev_dependencies:
   test: 1.16.0-nullsafety.9
   pedantic: 1.10.0-nullsafety.3
+  mockito: 4.1.1
+  fake_async: 1.1.0
 
   _fe_analyzer_shared: 12.0.0 # THIS LINE IS AUTOGENERATED - TO UPDATE USE "flutter update-packages --force-upgrade"
   analyzer: 0.40.6 # THIS LINE IS AUTOGENERATED - TO UPDATE USE "flutter update-packages --force-upgrade"
