+++ dev/benchmarks/metrics_center/test/flutter_test.dart
@@ -1,3 +1,7 @@
+// Copyright 2014 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
 import 'package:metrics_center/src/common.dart';
 import 'package:metrics_center/src/flutter.dart';
