+++ dev/benchmarks/metrics_center/test/legacy_flutter_test.dart
@@ -1,3 +1,7 @@
+// Copyright 2014 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
 import 'package:gcloud/src/datastore_impl.dart';
 import 'package:googleapis_auth/auth_io.dart';
 import 'package:metrics_center/src/common.dart';
